Fix polling and status reporting on deletion

+++ orc/internal/controllers/network/reconcile.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/gophercloud/gophercloud/v2/openstack/networking/v2/extensions/attributestags"
@@ -243,18 +244,13 @@ func (r *orcNetworkReconciler) reconcileDelete(ctx context.Context, orcObject *o
 		}
 		log.V(4).Info("Not deleting OpenStack resource due to policy", logPolicy...)
 	} else {
-		if len(orcObject.GetFinalizers()) > 1 {
-			log.V(4).Info("Deferring resource cleanup due to remaining external finalizers")
-			return ctrl.Result{}, nil
-		}
-
-		deleted, err := r.deleteResource(ctx, orcObject, addStatus)
+		deleted, requeue, err := r.deleteResource(ctx, log, orcObject, addStatus)
 		if err != nil {
 			return ctrl.Result{}, err
 		}
 
-		if !deleted {
-			return ctrl.Result{RequeueAfter: deletePollingPeriod}, nil
+		if deleted {
+			return ctrl.Result{RequeueAfter: requeue}, nil
 		}
 		log.V(4).Info("OpenStack resource is deleted")
 	}
@@ -266,10 +262,10 @@ func (r *orcNetworkReconciler) reconcileDelete(ctx context.Context, orcObject *o
 	return ctrl.Result{}, r.client.Patch(ctx, orcObject, ssa.ApplyConfigPatch(applyConfig), client.ForceOwnership, ssaFieldOwner(SSAFinalizerTxn))
 }
 
-func (r *orcNetworkReconciler) deleteResource(ctx context.Context, orcObject *orcv1alpha1.Network, addStatus func(updateStatusOpt)) (bool, error) {
+func (r *orcNetworkReconciler) deleteResource(ctx context.Context, log logr.Logger, orcObject *orcv1alpha1.Network, addStatus func(updateStatusOpt)) (bool, time.Duration, error) {
 	networkClient, err := r.getNetworkClient(ctx, orcObject)
 	if err != nil {
-		return false, err
+		return false, 0, err
 	}
 
 	if orcObject.Status.ID != nil {
@@ -283,18 +279,24 @@ func (r *orcNetworkReconciler) deleteResource(ctx context.Context, orcObject *or
 		switch {
 		case orcerrors.IsNotFound(err):
 			// Success!
-			return true, nil
+			return true, 0, nil
 
 		case err != nil:
-			return false, err
+			return false, 0, err
 
 		default:
 			addStatus(withResource(osResource))
+
+			if len(orcObject.GetFinalizers()) > 1 {
+				log.V(4).Info("Deferring resource cleanup due to remaining external finalizers")
+				return false, 0, nil
+			}
+
 			err := networkClient.DeleteNetwork(ctx, *orcObject.Status.ID).ExtractErr()
 			if err != nil {
-				return false, err
+				return false, 0, err
 			}
-			return false, nil
+			return false, deletePollingPeriod, nil
 		}
 	}
 
@@ -305,16 +307,16 @@ func (r *orcNetworkReconciler) deleteResource(ctx context.Context, orcObject *or
 	listOpts := listOptsFromCreation(orcObject)
 	osResource, err := getResourceFromList(ctx, listOpts, networkClient)
 	if err != nil {
-		return false, err
+		return false, 0, err
 	}
 
 	if osResource != nil {
 		addStatus(withResource(osResource))
-		return false, r.setStatusID(ctx, orcObject, osResource.ID)
+		return false, deletePollingPeriod, r.setStatusID(ctx, orcObject, osResource.ID)
 	}
 
 	// Didn't find an orphaned resource. Assume success.
-	return true, nil
+	return true, 0, nil
 }
 
 // getResourceName returns the name of the OpenStack resource we should use.
